requirements_win64.txt
altgraph==0.17
argcomplete==1.10.0
beautifulsoup4==4.8.0
cffi==1.14.4
chardet==3.0.4
cryptography==3.4.3
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
future==0.18.2
IMAPClient==2.1.0
lxml==4.6.2
olefile==0.46
pdfminer.six==20201018
pefile==2019.4.18
Pillow==8.1.0
pycparser==2.20
pycryptodome==3.10.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pypiwin32==223
python-pptx==0.6.18
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
six==1.12.0
sortedcontainers==2.3.0
soupsieve==2.1
SpeechRecognition==3.8.1
textract==1.6.3
tqdm==4.56.0
tzlocal==1.5.1
validate-docbr==1.8.2
xlrd==1.2.0
XlsxWriter==1.3.7
